strip out validate-mvh-account
