--- src/applications/validate-mhv-account/manifest.json
@@ -1,6 +1,0 @@
-{
-  "appName": "My Health Account Validation",
-  "entryFile": "./validate-mhv-account-entry",
-  "entryName": "my-health-account-validation",
-  "rootUrl": "/health-care/my-health-account-validation"
-}
